Update to 2019.3.3
This updates the version if IDEA that we're building against to 2019.3.3, and
includes some minor fixes needed to be used with this version.

Note that 2019.3.3 has a broken version of pydev/pydevd that prevents attaching
and debugging fusion scripts. It's possible to get it to work, but it's a bit
hacky. See https://github.com/JesusFreke/fusionIdea/issues/7#issuecomment-599841776
for more info.