Change how scripts are started and debugging is initiated in Fusion 360
This relies on an add-in running in Fusion 360 that exposes an HTTP server
that we connect to in order to start a script.

The current version of pydevd that is bundled with IDEA/PyCharm seems to
be a bit broken (see, e.g. https://youtrack.jetbrains.com/issue/PY-40359),
so we instead use a working version of pydevd that is bundled with the
plugin.