Sign the start script request
This allows the add-in to prompt for confirmation before running a script.

This gets rid of the python script that was used to communicate with the
add-in and trigger the script. Instead, this is all handled in java now.

This required some significant changes in the debug runner and related
classes to handle the "virtual" debugging process, now that we're not
launching an actual python process.

Currently, there is no difference between "running" and "debugging", and
the "attach to process" functionality does not work.