Add a dummy implementation of the GET_SMART_STEP_INTO_VARIANTS pdb pydevd command
This command was added recently to the IDEA/PyCharm forks of pydevd, but
the command id overlaps with another command (STEP_RETURN_MY_CODE) in the
main pydevd source.

This caused an issue where IDEA was issuing a GET_SMART_STEP_INTO_VARIANTS
command, but it as being interpreted as a STEP_RETURN_MY_CODE command, with
the result being that IDEA would freeze for a couple of minutes if you tried
to use "step into" while debugging a Fusion 360 script.

This change renumbers the STEP_RETURN_MY_CODE command (which seems to be
not used or implemented in IDEA), and adds a dummy no-op implementation
of the GET_SMART_STEP_INTO_VARIANTS command, which is enough to prevent
"step into" from freezing the IDE.