Improve how the process list is obtained
This fixes a problem that was occurring on start up, where an AVT event
from ProgressManager.runProcessWithProgressSynchronously showing a dialog
was causing an error due to not being allow to run during an "update" action.

The only solution I could find was to just not get the list of processes
while running on the dispatch thread.

I'm not positive this won't cause problems... but it seems to work so far.